# apps/api/app/modules/ai/imports/openapi_ai_imports.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Harry OS \u2014 AI Import Center",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/ai/imports/jobs": {
      "post": {
        "tags": [
          "ai",
          "ai-imports"
        ],
        "summary": "Create Import Session",
        "description": "Create a notebook import session (empty until images are uploaded).",
        "operationId": "create_import_session_api_v1_ai_imports_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportJobCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJobOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/ai/imports/jobs/{job_id}/pages": {
      "post": {
        "tags": [
          "ai",
          "ai-imports"
        ],
        "summary": "Upload Images",
        "description": "Upload one or more notebook page images to an import session.",
        "operationId": "upload_images_api_v1_ai_imports_jobs__job_id__pages_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_images_api_v1_ai_imports_jobs__job_id__pages_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportPageOut"
                  },
                  "title": "Response Upload Images Api V1 Ai Imports Jobs  Job Id  Pages Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "ai",
          "ai-imports"
        ],
        "summary": "Get Uploaded Images",
        "operationId": "get_uploaded_images_api_v1_ai_imports_jobs__job_id__pages_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ImportPageOut"
                  },
                  "title": "Response Get Uploaded Images Api V1 Ai Imports Jobs  Job Id  Pages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/imports/jobs/{job_id}/pages/{page_id}": {
      "delete": {
        "tags": [
          "ai",
          "ai-imports"
        ],
        "summary": "Delete Image",
        "operationId": "delete_image_api_v1_ai_imports_jobs__job_id__pages__page_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          },
          {
            "name": "page_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Page Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/imports/pages/{page_id}/file": {
      "get": {
        "tags": [
          "ai",
          "ai-imports"
        ],
        "summary": "Get Import Page File",
        "description": "Serve a stored notebook page image. Auth via Bearer or access_token query.",
        "operationId": "get_import_page_file_api_v1_ai_imports_pages__page_id__file_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Page Id"
            }
          },
          {
            "name": "access_token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional access token for <img src> (Bearer header preferred).",
              "title": "Access Token"
            },
            "description": "Optional access token for <img src> (Bearer header preferred)."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "image/*": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/imports/jobs/{job_id}": {
      "get": {
        "tags": [
          "ai",
          "ai-imports"
        ],
        "summary": "Get Import Status",
        "operationId": "get_import_status_api_v1_ai_imports_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportJobStatusOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/imports/jobs/{job_id}/preview": {
      "post": {
        "tags": [
          "ai",
          "ai-imports"
        ],
        "summary": "Generate Preview",
        "description": "Run extraction pipeline (scaffold) and return a reviewable journal draft.",
        "operationId": "generate_preview_api_v1_ai_imports_jobs__job_id__preview_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ImportPreviewRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportPreviewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ai/imports/jobs/{job_id}/commit": {
      "post": {
        "tags": [
          "ai",
          "ai-imports"
        ],
        "summary": "Save Journal",
        "description": "Commit the reviewed draft into Trading Journals (`source=native`).\n\nRejects with 409 if a journal already exists for the draft `journal_date`.",
        "operationId": "save_journal_api_v1_ai_imports_jobs__job_id__commit_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/ImportCommitRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportCommitOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConfidenceMap": {
        "properties": {
          "overall": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall"
          },
          "journal_date": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 1.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Journal Date"
          },
          "fields": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Fields"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "title": "ConfidenceMap"
      },
      "ImportCommitOut": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "journal_day_id": {
            "type": "string",
            "title": "Journal Day Id"
          },
          "journal_date": {
            "type": "string",
            "format": "date",
            "title": "Journal Date"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "review_status": {
            "type": "string",
            "title": "Review Status"
          },
          "trade_count": {
            "type": "integer",
            "title": "Trade Count"
          },
          "section_count": {
            "type": "integer",
            "title": "Section Count"
          },
          "attachment_count": {
            "type": "integer",
            "title": "Attachment Count"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "journal_day_id",
          "journal_date",
          "status",
          "review_status",
          "trade_count",
          "section_count",
          "attachment_count"
        ],
        "title": "ImportCommitOut"
      },
      "ImportCommitRequest": {
        "properties": {
          "draft": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JournalDraft-Input"
              },
              {
                "type": "null"
              }
            ]
          },
          "approve": {
            "type": "boolean",
            "title": "Approve",
            "default": true
          }
        },
        "type": "object",
        "title": "ImportCommitRequest",
        "description": "Save Journal \u2014 optional draft override; otherwise uses stored preview draft."
      },
...
      "ImportJobCreate": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "notebook_label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Notebook Label"
          },
          "detected_journal_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Journal Date"
          }
        },
        "type": "object",
        "title": "ImportJobCreate"
      },
      "ImportJobOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "notebook_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notebook Label"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "review_status": {
            "type": "string",
            "title": "Review Status"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "overall_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Confidence"
          },
          "draft_version": {
            "type": "integer",
            "title": "Draft Version"
          },
          "detected_journal_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Journal Date"
          },
          "committed_journal_day_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Committed Journal Day Id"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "notebook_label",
          "status",
          "review_status",
          "current_stage",
          "page_count",
          "overall_confidence",
          "draft_version",
          "detected_journal_date",
          "committed_journal_day_id",
          "model_id",
          "prompt_version",
          "error_code",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "ImportJobOut"
      },
      "ImportJobStatusOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "notebook_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notebook Label"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "review_status": {
            "type": "string",
            "title": "Review Status"
          },
          "current_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Stage"
          },
          "page_count": {
            "type": "integer",
            "title": "Page Count"
          },
          "overall_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overall Confidence"
          },
          "draft_version": {
            "type": "integer",
            "title": "Draft Version"
          },
          "detected_journal_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detected Journal Date"
          },
          "committed_journal_day_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Committed Journal Day Id"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "has_draft": {
            "type": "boolean",
            "title": "Has Draft",
            "default": false
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConfidenceMap"
              },
              {
                "type": "null"
              }
            ]
          },
          "pages": {
            "items": {
              "$ref": "#/components/schemas/ImportPageOut"
            },
            "type": "array",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "notebook_label",
          "status",
          "review_status",
          "current_stage",
          "page_count",
          "overall_confidence",
          "draft_version",
          "detected_journal_date",
          "committed_journal_day_id",
          "model_id",
          "prompt_version",
          "error_code",
          "error_message",
          "created_at",
          "updated_at"
        ],
        "title": "ImportJobStatusOut",
        "description": "Status poll payload \u2014 includes draft presence + confidence map."
      },
      "ImportPageOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "page_index": {
            "type": "integer",
            "title": "Page Index"
          },
          "original_file_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original File Name"
          },
          "mime_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime Type"
          },
          "byte_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Byte Size"
          },
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "quality_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quality Score"
          },
          "ocr_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Confidence"
          },
          "has_ocr_transcript": {
            "type": "boolean",
            "title": "Has Ocr Transcript",
            "default": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "job_id",
          "page_index",
          "original_file_name",
          "mime_type",
          "byte_size",
          "checksum",
          "status",
          "quality_score",
          "ocr_confidence",
          "created_at",
          "updated_at"
        ],
        "title": "ImportPageOut"
      },
      "ImportPreviewOut": {
        "properties": {
          "job": {
            "$ref": "#/components/schemas/ImportJobOut"
          },
          "draft": {
            "$ref": "#/components/schemas/JournalDraft-Output"
          },
          "confidence": {
            "$ref": "#/components/schemas/ConfidenceMap"
          },
          "draft_version": {
            "type": "integer",
            "title": "Draft Version"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "job",
          "draft",
          "confidence",
          "draft_version"
        ],
        "title": "ImportPreviewOut"
      },
      "ImportPreviewRequest": {
        "properties": {
          "journal_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Journal Date"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "title": "ImportPreviewRequest",
        "description": "Optional overrides when generating a preview draft."
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}